feat: ブックマークと履歴のサービスおよび変換機能を追加し、タブ関連のユーティリティを実装 (#26)
