+++ src/services/bookmark/types.ts
