--- src/services/tab/tab.converter.ts
@@ -1,35 +1,0 @@
-import { ResultType } from "@/types/result";
-import { calcMatchRateResult } from "@/utils/match";
-import * as Type from "./tab.types";
-
-/**
- * タブデータの変換を担当するクラス
- */
-export class TabConverter {
-  /**
-   * Chrome APIのタブデータを内部のタブ形式に変換する
-   */
-  static convertToTabs(
-    chromeTabs: chrome.tabs.Tab[],
-    query?: string
-  ): Type.Tab[] {
-    return chromeTabs.map((chromeTab) => {
-      const tab: Type.Tab = {
-        type: ResultType.Tab,
-        id: chromeTab.id || 0,
-        title: chromeTab.title || "",
-        url: chromeTab.url || "",
-        icon: chromeTab.favIconUrl || "",
-        active: chromeTab.active || false,
-        lastAccessed: chromeTab.lastAccessed || 0,
-        windowId: chromeTab.windowId || 0,
-        currentWindow: false, // TODO: 現在のウィンドウかどうかの判定ロジックが必要な場合
-        match: query
-          ? calcMatchRateResult(query, chromeTab.title, chromeTab.url)
-          : 0,
-      };
-
-      return tab;
-    });
-  }
-}
